Some tests for paired BAM reader
